refactor: Move initApp function to its own file
